+++ apps/docs/src/app/core/api-files.ts
@@ -130,7 +130,7 @@ export const API_FILES = {
         'ListSecondaryDirective',
         'ListGroupHeaderDirective',
         'ListIconDirective',
-        'ListFooterDirective',
+        'ListFooterDirective'
     ],
     loadingSpinner: [
         'LoadingSpinnerComponent'