+++ apps/docs/src/app/core/component-docs/multi-input/examples/multi-input-compact-example/multi-input-compact-example.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'fd-multi-input-compact-example',
+    templateUrl: './multi-input-compact-example.component.html'
+})
+export class MultiInputCompactExampleComponent {
+
+    selected = ['Banana'];
+
+}