+++ apps/docs/src/app/core/component-docs/shellbar/shellbar-docs.component.ts
@@ -4,8 +4,6 @@ import * as shellbarBasicHTMLSrc from '!raw-loader!./examples/shellbar-basic-exa
 import * as shellbarBasicTSSrc from '!raw-loader!./examples/shellbar-basic-example.component.ts';
 import * as shellbarCollapsibleHTMLSrc from '!raw-loader!./examples/shellbar-collapsible-example.component.html';
 import * as shellbarCollapsibleTSSrc from '!raw-loader!./examples/shellbar-collapsible-example.component.ts';
-import * as advancedShellbarTs from '!raw-loader!./examples/shellbar-advanced/shellbar-advanced-example.component.ts';
-import * as advancedShellbarHtml from '!raw-loader!./examples/shellbar-advanced/shellbar-advanced-example.component.html';
 import * as sideNavShellbarHtml from '!raw-loader!./examples/shellbar-side-nav/shellbar-side-nav-example.component.html';
 import * as sideNavShellbarTs from '!raw-loader!./examples/shellbar-side-nav/shellbar-side-nav-example.component.ts';
 import { ExampleFile } from '../../../documentation/core-helpers/code-example/example-file';
@@ -40,20 +38,6 @@ export class ShellbarDocsComponent {
             component: 'ShellbarCollapsibleExampleComponent',
             code: shellbarCollapsibleTSSrc,
             fileName: 'shellbar-collapsible-example'
-        }
-    ];
-
-    advancedShellbar: ExampleFile[] = [
-        {
-            language: 'html',
-            code: advancedShellbarHtml,
-            fileName: 'shellbar-advanced-example'
-        },
-        {
-            language: 'typescript',
-            component: 'ShellbarAdvancedExampleComponent',
-            code: advancedShellbarTs,
-            fileName: 'shellbar-advanced-example'
         }
     ];