+++ libs/core/src/lib/checkbox/checkbox/checkbox.component.ts
@@ -85,8 +85,10 @@ export class CheckboxComponent implements ControlValueAccessor {
     /** @hidden Reference to callback provided by FormControl.*/
     public onValueChange = (newValue) => {};
 
+    /** @hidden */
     constructor(
-        private _changeDetectorRef: ChangeDetectorRef
+        private _changeDetectorRef: ChangeDetectorRef,
+        public elementRef: ElementRef
     ) {}
 
     /** @hidden Used to define if control is in 'indeterminate' state.*/