+++ libs/core/src/lib/list/list-item.directive.ts
@@ -1,4 +1,5 @@
-import { Directive, HostBinding, Input } from '@angular/core';
+import { ContentChild, Directive, ElementRef, HostBinding, Input } from '@angular/core';
+import { CheckboxComponent } from '../checkbox/checkbox/checkbox.component';
 
 /**
  * The component that represents a list item.
@@ -27,4 +28,28 @@ export class ListItemDirective {
     @HostBinding('class.is-selected')
     selected: boolean = false;
 
+    /** @hidden */
+    @ContentChild(CheckboxComponent, {static: false})
+    checkboxComponent: CheckboxComponent;
+
+    /** @hidden */
+    constructor (public itemEl: ElementRef) {}
+
+    /** @hidden */
+    public focus(): void {
+        if (this.checkboxComponent) { // if there is a checkbox in this list item, we want to focus its input label
+            this.checkboxComponent.inputLabel.nativeElement.focus();
+        } else {
+            this.itemEl.nativeElement.focus();
+        }
+    }
+
+    /** @hidden */
+    public click(): void {
+        if (this.checkboxComponent) { // if there is a checkbox in this list item, we want to click its input label
+            this.checkboxComponent.inputLabel.nativeElement.click();
+        } else {
+            this.itemEl.nativeElement.click();
+        }
+    }
 }